+++ core/client/headers.go
@@ -1,6 +1,7 @@
 package client
 
 import (
+	"fmt"
 	"time"
 )
 
@@ -13,6 +14,7 @@ const (
 
 func (c *Client) buildHeaders() map[string]string {
 	date := time.Now()
+	fmt.Println("date", date.String())
 	headers := map[string]string{
 		DateHeader:     date.Format(dateFormat),
 		AuthHeader:     c.config.AuthHeader(date),