+++ core/client/response.go
@@ -1,6 +1,9 @@
 package client
 
-import "encoding/json"
+import (
+	"encoding/json"
+	"github.com/go-resty/resty/v2"
+)
 
 type Response[T any] struct {
 	Payload []byte
@@ -8,10 +11,12 @@ type Response[T any] struct {
 	Error   error
 }
 
-func NewResponse[T any](payload []byte, err error) Response[T] {
+func NewResponse[T any](res *resty.Response, err error) Response[T] {
 	if err != nil {
 		return ErrorResponse[T](err)
 	}
+
+	payload := res.Body()
 
 	var resp Response[T]
 	resp.Payload = payload