+++ license_client/airgap.go
@@ -1,14 +1,17 @@
 package license_client
 
 import (
+	"crypto/ecdsa"
+	"crypto/x509"
 	"encoding/base64"
 	"encoding/json"
+	"encoding/pem"
 	"fmt"
 	. "gitlab.com/l3178/sdk-go/core/models"
 	"gitlab.com/l3178/sdk-go/license_client/airgap"
 )
 
-func (c *LicenseClient) AirgapInitialization(licenseKey, signingKey string) (string, error) {
+func (c *AirgapClient) AirgapInitialization(licenseKey, signingKey string) (string, error) {
 	decoded, err := base64.StdEncoding.DecodeString(signingKey)
 	if err != nil {
 		return "", fmt.Errorf("invalid initialization code: %w", err)
@@ -21,7 +24,7 @@ func (c *LicenseClient) AirgapInitialization(licenseKey, signingKey string) (str
 	return airgap.GenerateSignature(privateKey, signingString)
 }
 
-func (c *LicenseClient) AirgapActivation(licensePolicy, confirmationCode string, policyId int64) (LicenseResponse, error) {
+func (c *AirgapClient) AirgapActivation(licensePolicy, confirmationCode string, policyId int64) (LicenseResponse, error) {
 	lcsJson, err := base64.StdEncoding.DecodeString(licensePolicy)
 	if err != nil {
 		return LicenseResponse{}, fmt.Errorf("invalid license policy: %w", err)
@@ -30,6 +33,8 @@ func (c *LicenseClient) AirgapActivation(licensePolicy, confirmationCode string,
 	if confirmationCode == "" {
 		return LicenseResponse{}, fmt.Errorf("invalid confirmation code")
 	}
+
+	c.verifyConfirmationCode(confirmationCode)
 
 	var license LicenseResponse
 
@@ -43,4 +48,18 @@ func (c *LicenseClient) AirgapActivation(licensePolicy, confirmationCode string,
 	}
 
 	return license, nil
+}
+
+func (c *AirgapClient) verifyConfirmationCode(confirmationCode string) bool {
+	key := c.publicKey()
+
+	return ecdsa.VerifyASN1(key, []byte("LicenseSpring"), []byte(confirmationCode))
+}
+
+func (c *AirgapClient) publicKey() *ecdsa.PublicKey {
+	fullKey := fmt.Sprintf("-----BEGIN PUBLIC KEY-----\n%s\n-----END PUBLIC KEY-----", c.PublicKey)
+
+	decoded, _ := pem.Decode([]byte(fullKey))
+	publicKey, _ := x509.ParsePKIXPublicKey(decoded.Bytes)
+	return publicKey.(*ecdsa.PublicKey)
 }