+++ changelog.md
@@ -1,3 +1,20 @@
+v1.0.13. - 20.6.2023.
+- General improvements
+- More Management API tests
+
+v1.0.12. - 16.6.2023.
+- General fixes and improvements
+- License client tests
+
+v1.0.11. - 6.6.2023.
+- Management client tests
+
+v1.0.10. - 6.6.2023.
+- Management client updates
+
+v1.0.9. - 8.5.2023.
+- API response struct change
+
 v1.0.8 - 3.2.2023.
 - Airgap implementation